bot/requirements.txt
# Django and Django REST Framework
Django==5.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3
django-debug-toolbar==4.2.0
django-rest-framework-simplejwt==5.4.0

# Database
psycopg2-binary==2.9.9

# Environment and Configuration
python-dotenv==1.0.0
python-decouple==3.8

# AI and Machine Learning
openai==1.3.7
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10

# Vector Database
pinecone-client==2.2.4

# Web Scraping and Parsing
beautifulsoup4==4.12.2
aiohttp==3.9.1
requests==2.31.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
docx2txt==0.8

# Payment Processing
stripe==10.1.0
paypal-checkout-sdk==1.0.1

# Authentication and Security
PyJWT==2.8.0

# Background Tasks and Caching
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1

# Image Processing
Pillow==10.1.0

# Web Server
gunicorn==21.2.0
whitenoise==6.6.0

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0

# Utilities
python-dateutil==2.8.2 